requirements.txt
fastapi
psycopg[binary]
dotenv
uvicorn
pyYAML
requests
